I got this result when running tests for node insertion using the graph api and federation api (just for test purposes):
Graph api:
Nodes   Seconds
100     3.6
500     5.8
1000    7.6
2000    10.96
4000    18
8410    37.9 ~ 2000 node errors

Federation api (using project node):
Nodes   Seconds
100     7.2
500     10
1000    14.4
2000    24
4000    37 // here I needed to upgrade jvm memory
8407    83.6

It seems that the graph api is more fast. Maybe the strategy of storing all nodes on memory and put on jcr only on save used on federation api should cause some performance lack. The strategy of saving each node on jcr as soon as possible used on graph api seems to be more fast.

But, on using links, I just tested that on graph api. It required 1060 seconds to insert the same nodes as I did before and also about 11000 links. I think this time should be better...

git-svn-id: https://openspotlight.dev.java.net/svn/openspotlight/trunk/source/osl@137 65b4daa9-cd5d-0410-b3a2-f97e3e1d3c1c